ci: message handler redesigned #70
