+++ src/consensus/src/lib.rs
@@ -6,3 +6,5 @@ extern crate log;
 
 pub mod consensus_config;
 pub mod consensus_interface;
+pub mod consensus_message_sender;
+pub mod consensus_messages;
