+++ src/controllers/Cargo.toml
@@ -8,7 +8,7 @@ license = "MIT OR Apache-2.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-p2plib = { path = "../p2plib" }
+message_handler = { path = "../message_handler" }
 services = { path = "../services" }
 utils = { path = "../utils" }
