+++ src/message_handler/Cargo.toml
@@ -1,19 +1,18 @@
 [package]
-name = "p2plib"
+name = "message_handler"
 version = "0.1.0"
-authors = ["Pankajm"]
+authors = ["gaurav agarwal <gaurav.agarwal@talentica.com>"]
 edition = "2018"
-license = "MIT OR Apache-2.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 utils = { path = "../utils" }
 sdk = { path = "../sdk" }
-message_handler = { path = "../message_handler" }
+schema = { path = "../schema" }
 
 futures = "^0.3"
 libp2p = "^0.15"
 serde = { version = "1.0", features = ["derive"] }
-serde_cbor = "^0.11"
+lazy_static = "^1.4"
 log = "0.4.8"
