+++ src/message_handler/src/lib.rs
@@ -1,0 +1,9 @@
+#[macro_use]
+extern crate lazy_static;
+
+#[macro_use]
+extern crate log;
+
+pub mod message_sender;
+pub mod messages;
+pub mod node_messages;
