+++ src/node/Cargo.toml
@@ -15,6 +15,7 @@ p2plib = { path = "../p2plib" }
 schema = { path = "../schema" }
 utils = { path = "../utils" }
 sdk = { path = "../sdk"}
+message_handler = { path = "../message_handler" }
 
 exonum-merkledb = "^1.0.0-rc.3"
 exonum-crypto = "^1.0.0-rc.3"
