--- src/schema/src/transaction.rs
@@ -1,2 +1,0 @@
-// pub use doc_app::{state::State, transaction::SignedTransaction};
-// pub use exonum_merkledb::ObjectHash;
