+++ src/sdk/src/constants.rs
@@ -2,3 +2,4 @@
 pub const NODE: &'static str = "node";
 pub const CONSENSUS: &'static str = "consensus";
 pub const P2P: &'static str = "p2p";
+pub const NODE_MSG_TOPIC_STR: &'static [&'static str] = &["SignedTransaction", "SignedBlock"];
