+++ src/sdk/src/lib.rs
@@ -1,6 +1,8 @@
 #[macro_use]
 extern crate exonum_derive;
 
+pub mod constants;
+pub mod message_traits;
 pub mod signed_transaction;
 pub mod state;
 pub mod traits;
