+++ src/services/Cargo.toml
@@ -9,7 +9,7 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 db_service = { path = "../db_service"}
-p2plib = { path = "../p2plib" }
+message_handler = { path = "../message_handler" }
 schema = { path = "../schema" }
 utils = { path = "../utils" }
 
@@ -17,13 +17,6 @@ exonum-merkledb = "^1.0.0-rc.3"
 exonum-crypto = "^1.0.0-rc.3"
 exonum-derive = "^1.0.0-rc.3"
 
-hex = "^0.4"
-serde = { version = "1.0", features = ["derive"] }
-serde_cbor = "^0.10"
-libp2p = "^0.15"
 futures = "0.3.1"
-async-std = "^1.4"
-lazy_static = "^1.4"
 actix-web = "2.0"
-actix-rt = "1.0"
 log = "0.4.8"
