+++ src/services/src/client_services.rs
@@ -3,8 +3,8 @@ use db_service::db_layer::snapshot_db;
 use db_service::db_snapshot_ref::SchemaSnap;
 use exonum_crypto::Hash;
 use futures::channel::mpsc::*;
-use p2plib::message_sender::MessageSender;
-use p2plib::messages::MessageTypes;
+use message_handler::message_sender::MessageSender;
+use message_handler::messages::MessageTypes;
 use schema::signed_transaction::SignedTransaction;
 use schema::transaction_pool::{TxnPool, TxnPoolKeyType, POOL};
 use utils::serializer::{deserialize, serialize};
