+++ src/services/src/lib.rs
@@ -2,11 +2,3 @@
 extern crate log;
 
 pub mod client_services;
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
